ahead of changes

+++ createanimal.py
@@ -23,10 +23,6 @@ class CreatePet(Animal):
         return cat1
 
 
-dog1 = CreatePet.dog()
-cat1 = CreatePet.cat()
-
-
 class DogActions(CreatePet):
 
     def action(self):
@@ -54,34 +50,40 @@ class DogActions(CreatePet):
                 print("check status")
                 print("kill dog")
                 print("show cat")
-        else:
-            pass
+            else:
+                pass
 
+
+cat1 = CreatePet.cat()
+
+dog1 = CreatePet.dog()
 
 
 class CatActions(CreatePet):
-
     def action(self):
-        player_action = self
-        if player_action == "name cat":
-            cat1.change_name()
-        if player_action == "feed cat":
-            cat1.feed()
-        if player_action == "play with cat":
-            cat1.play()
-        if player_action == "put cat to bed":
-            cat1.sleep()
-        if player_action == "check status":
-            cat1.check_status()
-        if player_action == "kill cat":
-            cat1.kill()
-        if player_action == "show cat":
-            cat1.meow()
-        if player_action == "help":
-            print("List of commands as follows:")
-            print("feed cat")
-            print("play with cat")
-            print("put cat to bed")
-            print("check status")
-            print("kill cat")
-            print("show cat")
+        if cat1.alive == True:
+            player_action = self
+            if player_action == "name cat":
+                cat1.change_name()
+            if player_action == "feed cat":
+                cat1.feed()
+            if player_action == "play with cat":
+                cat1.play()
+            if player_action == "put cat to bed":
+                cat1.sleep()
+            if player_action == "check status":
+                cat1.check_status()
+            if player_action == "kill cat":
+                cat1.kill()
+            if player_action == "show cat":
+                cat1.meow()
+            if player_action == "help":
+                print("List of commands as follows:")
+                print("feed cat")
+                print("play with cat")
+                print("put cat to bed")
+                print("check status")
+                print("kill cat")
+                print("show cat")
+            else:
+                pass
